refactor(turn): extract combat and end-of-turn phase runners
- Extract combat phase logic to combat_phase_helpers.dart (runOneLandBattle,
  runOneNavalBattle, _runCombatPhase helpers)
- Extract end-of-turn logic to end_of_turn_resolver.dart
- Slim turn_resolver.dart; remove unused combat imports
- Add conflict_detection import to combat_phase_helpers for BattleContext

Made-with: Cursor
