+++ packages/colonizethis_logic/lib/src/turn/combat_phase_helpers.dart
@@ -1,0 +1,117 @@
+// Helpers for the combat phase: apply one land battle (quick or auto-resolve), evidence, dialogue.
+// SPEC/program/turn-resolution-phase-details.md § Combat. Called from turn_resolver._runCombatPhase.
+
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../combat/combat_resolver.dart';
+import '../combat/conflict_detection.dart';
+import '../combat/quick_battle_input_builder.dart';
+import '../combat/quick_battle_resolver.dart';
+import '../constants.dart';
+import '../dossier/evidence_rules.dart';
+import '../dossier/event_dialogue.dart';
+
+/// Runs one land battle: applies result (quick battle or auto-resolve), evidence, and dialogue.
+Game runOneLandBattle(
+  Game state,
+  BattleContext ctx,
+  CombatMode mode,
+  Map<String, double> feedingCoverageByPlayerId,
+  int turn,
+  int battleIndex,
+  int seed, {
+  void Function(DialogueEvent)? onDialogue,
+}) {
+  if (mode == CombatMode.quickBattle) {
+    final input = buildQuickBattleInput(state, ctx,
+        seed: state.worldState.turnState.turnNumber);
+    final qbResult = resolveQuickBattle(input);
+    state = applyQuickBattleResultToGame(state, ctx, qbResult);
+    if (qbResult.winner == QuickBattleWinner.attacker &&
+        qbResult.provinceFlips &&
+        ctx.attackers.isNotEmpty) {
+      final victorId = ctx.attackers.first.factionId;
+      final evidence = evidenceForLandBattleVictory(
+          state, victorId, ctx.defenderFactionId, turn);
+      if (evidence.isNotEmpty) {
+        state = state.copyWith(dossierEvidenceEntries: [
+          ...state.dossierEvidenceEntries,
+          ...evidence
+        ]);
+      }
+      _emitLandBattleDialogue(
+          state, victorId, ctx.defenderFactionId, ctx.provinceId, turn,
+          battleIndex, seed, onDialogue);
+    } else {
+      final victorId = qbResult.winner == QuickBattleWinner.defender
+          ? ctx.defenderFactionId
+          : (qbResult.provinceFlips && ctx.attackers.isNotEmpty
+              ? ctx.attackers.first.factionId
+              : null);
+      final loserId =
+          victorId == ctx.defenderFactionId && ctx.attackers.isNotEmpty
+              ? ctx.attackers.first.factionId
+              : (victorId != null ? ctx.defenderFactionId : null);
+      if (victorId != null && loserId != null) {
+        _emitLandBattleDialogue(
+            state, victorId, loserId, ctx.provinceId, turn, battleIndex, seed,
+            onDialogue);
+      }
+    }
+  } else {
+    state = resolveBattleContext(
+      state,
+      ctx,
+      feedingCoverageByPlayerId: feedingCoverageByPlayerId,
+    );
+    final region = ctx.regionId == kRegionOldWorld
+        ? state.worldState.oldWorld
+        : state.worldState.newWorld;
+    final province =
+        region.provinces.where((p) => p.id == ctx.provinceId).firstOrNull;
+    final victorId = province?.ownerId;
+    if (victorId != null && victorId != ctx.defenderFactionId) {
+      final evidence = evidenceForLandBattleVictory(
+          state, victorId, ctx.defenderFactionId, turn);
+      if (evidence.isNotEmpty) {
+        state = state.copyWith(dossierEvidenceEntries: [
+          ...state.dossierEvidenceEntries,
+          ...evidence
+        ]);
+      }
+    }
+    final effectiveVictorId = victorId ?? ctx.defenderFactionId;
+    final effectiveLoserId =
+        victorId == ctx.defenderFactionId && ctx.attackers.isNotEmpty
+            ? ctx.attackers.first.factionId
+            : ctx.defenderFactionId;
+    _emitLandBattleDialogue(state, effectiveVictorId, effectiveLoserId,
+        ctx.provinceId, turn, battleIndex, seed, onDialogue);
+  }
+  return state;
+}
+
+void _emitLandBattleDialogue(
+  Game state,
+  String victorId,
+  String loserId,
+  String provinceId,
+  int turn,
+  int battleIndex,
+  int seed,
+  void Function(DialogueEvent)? onDialogue,
+) {
+  if (onDialogue == null) return;
+  final dialogueSeed = (seed ^ (battleIndex * 0x9E3779B1)) & 0x7fffffff;
+  final events = dialogueEventsForLandBattleResult(
+    state,
+    victorId,
+    loserId,
+    provinceId,
+    turn,
+    dialogueSeed,
+  );
+  if (events.isNotEmpty) {
+    for (final e in events) onDialogue(e);
+  }
+}
